Prepared bin directory and gruntfile for switch to minimist.
